- double check the total/tax calculation - Delivery Fee shows the decimal points. - Change the confirm button says "I agree" to "Yes"  in checkout screen - remove space after the dollar signs on the checkout screen - make the banner a little thinner - banner doesn't bounces anymore

+++ app/src/main/java/com/bentonow/bentonow/model/order/OrderDetails.java
@@ -4,9 +4,9 @@ public class OrderDetails {
     public OrderAddress address = new OrderAddress();
     public OrderLocation coords = new OrderLocation();
     public int coupon_discount_cents = 0;
-    public int tip_percentage = 15;
-    public long tax_cents = 0;
-    public int tip_cents = 0;
-    public long total_cents = 0;
-    public long delivery_price = 0;
+    public double tip_percentage = 15;
+    public double tax_cents = 0;
+    public double tip_cents = 0;
+    public double total_cents = 0;
+    public double delivery_price = 0;
 }
